菜单渲染因错误的使用useEffect导致的打开和默认选中选项不可用问题修复.这块逻辑不适合用euseEffect来处理,所以使用useMemo来做了性能优化
