+++ webpack.config.js
@@ -8,6 +8,7 @@ module.exports = {
       "@": path.resolve(__dirname, "src"),
       "@configs": path.resolve(__dirname, "src/configs"),
       "@hooks": path.resolve(__dirname, "src/hooks"),
+      "@utils": path.resolve(__dirname, "src/utils"),
       "@components": path.resolve(__dirname, "src/components"),
       "@pages": path.resolve(__dirname, "src/pages"),
       "@routes": path.resolve(__dirname, "src/routes"),
